2006-02-27  Eric Seidel  <eseidel@apple.com>
        Reviewed by darin.

        Build fix: cast to unsigned from size_t.

        * platform/StreamingTextDecoder.cpp:


Canonical link: https://commits.webkit.org/10993@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@13013 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/platform/StreamingTextDecoder.cpp
@@ -95,7 +95,7 @@ QString StreamingTextDecoder::convertUTF16(const unsigned char *s, int length)
     
     while (len > 1) {
         UChar buffer[ConversionBufferSize];
-        int runLength = min(len / 2, sizeof(buffer) / sizeof(buffer[0]));
+        int runLength = min(len / 2, (unsigned)(sizeof(buffer) / sizeof(buffer[0])));
         int bufferLength = 0;
         if (m_littleEndian) {
             for (int i = 0; i < runLength; ++i) {
